// src/interfaces/user.ts
export interface User {
    userId: string; // Generated by server using UUID
    username: string;
    solanaWalletAddress: string;
    password: string; // Should be hashed in production
    email: string;
    firstName: string;
    lastName: string;
    birthdate: string;
    age: number;
    country: string;
    mobileNumber: string;
    twitterHandle: string;
    discordId: string;
    telegramId: string;
    credit: number;
  }